api/requirements.txt
fastapi 
uvicorn
ultralytics
python-dotenv
boto3 
sqlalchemy 
psycopg2-binary 
Pillow